+++ src/com/amplitude/api/Identify.java
@@ -270,6 +270,79 @@ public class Identify {
     }
 
 
+    // PREPEND
+    public Identify prepend(String property, boolean value) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, value);
+        return this;
+    }
+
+    public Identify prepend(String property, double value) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, value);
+        return this;
+    }
+
+    public Identify prepend(String property, float value) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, value);
+        return this;
+    }
+
+    public Identify prepend(String property, int value) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, value);
+        return this;
+    }
+
+    public Identify prepend(String property, long value) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, value);
+        return this;
+    }
+
+    public Identify prepend(String property, String value) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, value);
+        return this;
+    }
+
+    public Identify prepend(String property, JSONArray values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, values);
+        return this;
+    }
+
+    // Server-side we flatten dictionaries and apply prepend to each flattened property
+    public Identify prepend(String property, JSONObject values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, values);
+        return this;
+    }
+
+    public Identify prepend(String property, boolean[] values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, booleanArrayToJSONArray(values));
+        return this;
+    }
+
+    public Identify prepend(String property, double[] values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, doubleArrayToJSONArray(values));
+        return this;
+    }
+
+    public Identify prepend(String property, float[] values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, floatArrayToJSONArray(values));
+        return this;
+    }
+
+    public Identify prepend(String property, int[] values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, intArrayToJSONArray(values));
+        return this;
+    }
+
+    public Identify prepend(String property, long[] values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, longArrayToJSONArray(values));
+        return this;
+    }
+
+    public Identify prepend(String property, String[] values) {
+        addToUserProperties(Constants.AMP_OP_PREPEND, property, stringArrayToJSONArray(values));
+        return this;
+    }
+
+
     // UNSET
     public Identify unset(String property) {
         addToUserProperties(Constants.AMP_OP_UNSET, property, "-");